fix bunch of bugs

+++ backend/sidhandler.py
@@ -95,7 +95,7 @@ class SIDCache:
 	def allauth(self):
 		return [sid.uuid for sid in self.sids if sid.authstate]
 	def cutauths(self):
-		return [sid.uuid[:len(sid.uuid)/2] for sid in self.sids if sid.authstate]
+		return [sid.uuid[:int(len(sid.uuid)/2)] for sid in self.sids if sid.authstate]
 	def allnonauth(self):
 		return [sid.uuid for sid in self.sids if not sid.authstate]
 	def _isin(self, uuid):
